pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "python-pkg"
description = "An alternative to Python's `pip`, with inspiration for Julia's `Pkg`!"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
version = "0.1.0.dev"

keywords = [
  "package-manager", "pip", "metadata", "packages",
]

authors = [
  { name = "Joe Carpinelli", email = "jdcarpinelli@gmail.com" },
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dependencies = [
  "typer",
  "rich",
  "requests",
  "packaging",
  "module-hygiene",
]

[project.urls]
Documentation = "https://github.com/cadojo/python-pkg#readme"
Issues = "https://github.com/cadojo/python-pkg/issues"
Source = "https://github.com/cadojo/python-pkg"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=pkg --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
